Removed different PagingEntitiesTest and just consolidated tests into PagingResourceIT. Added helper methods and removed a lot of duplicated code. Added queryParameters to getNextPage method since we can do queryGets. Added a note to have the collection remember its own name
