+++ stack/rest/src/test/java/org/apache/usergrid/rest/test/resource2point0/endpoints/CollectionEndpoint.java
@@ -115,9 +115,13 @@ public class CollectionEndpoint extends NamedResource {
      * }     *
      * usersCollection = app.collections("users").getNextPage(usersCollection.cursor);
      */
-    public Collection getNextPage(Collection collection, final boolean useToken) {
+    //TODO: add queryParameters here
+    public Collection getNextPage(Collection collection, QueryParameters passedParameters ,final boolean useToken) {
         WebResource resource = getResource(useToken);
-        QueryParameters queryParameters = new QueryParameters();
+        QueryParameters queryParameters = passedParameters;
+        if( queryParameters == null){
+            queryParameters = new QueryParameters();
+        }
 
         queryParameters.setCursor(collection.getCursor());
         resource = addParametersToResource(resource, queryParameters);
